refactor(ui): address review round-3 findings on ModelIndicator
- Fix switchProviderModel rollback: capture previousModelId before
  calling updateProviderModel so the rollback target is the original
  value, not the already-persisted new one
- Change DropdownMenuItem from onClick to onSelect + preventDefault so
  keyboard activation (Space/Enter) works and we control menu-close
  timing after the async update resolves
- Fix otherProviders separator: always render ProviderSubMenu entries
  when otherProviders.length > 0; only render the separator when
  preceding content exists (hasModel || siblingModels.length > 0)
- Split ModelIndicator.unit.test.tsx (258 lines) into focused files:
  - ModelIndicator.same-provider.unit.test.tsx (US1 + US3 + running states)
  - ModelIndicator.cross-provider.unit.test.tsx (US2)
  - ModelIndicator.test-utils.ts (shared baseSettings helper)
  Use vi.hoisted() for mock variables to avoid vi.mock hoisting fragility
